+++ src/devpi_tokens/views.py
@@ -8,8 +8,10 @@ from pyramid.view import view_config
     permission="user_modify")
 def user_token_create(context, request):
     tu = request.devpi_token_utility
+    restrictions = [
+        tu.get_expires_restriction(request)]
     result = dict(
-        token=tu.new_token(context.user))
+        token=tu.new_token(context.user, restrictions))
     apireturn(200, type="token-info", result=result)